requirements.txt
# FastAPI e servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn==21.2.0
python-multipart==0.0.6

# Banco de dados
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.0.1

# OpenAI / Agno (quando disponível)
openai>=1.54.0

# Stripe
stripe==7.0.0

# HTTP Client
httpx>=0.27.0

# Utilitários
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Logging
python-json-logger==2.0.7

# Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# RAG - Base de Conhecimento
pypdf==3.17.0
tiktoken==0.5.2
pgvector==0.2.4

# MCP Server
mcp>=0.9.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoramento (opcional)
psutil==5.9.6